Improve Error Message for Delegate Methods with State Issues
Summary:
## This Diff

One case that’s missing is: handling improper state issues within delegate methods.  That currently has the error:
```
Argument at index 1 is not a valid parameter, should be one of the following: Prop T somePropName. TreeProp T someTreePropName. StateValue<T> stateName, where a state param with type T and name stateName is declared elsewhere in the spec. InjectProp T someInjectPropName.
```

So, let’s cover this last case and ensure that we provide a proper error!

## More Context

In looking at the history, we purposefully moved delegate method validation to happen earlier than the other generic spec validation portions.  This was to fix some other errors that were less clear if the generic spec model validation came first.

So, all that said, we shold just handle this potential covariant state issue in the `DelegateMethodValidation` as reordering this check to come first was purposeful.

When testing, this issue *only* applies to `StateValue` and not to cached values (those ones stll have the normal error message), so this diff is foxused only on resolving the issue for state.

## What’s Left

Fundamentally, we could take it a step further and attempt to leave covariance allowed and just update the code generator.  In essence, we would need to process the state values that we pull (for method params) to remove covariance when populating the spec model.

If this were 2 years ago, I’d likely consider doing that next step, but it’s not.  As we’re moving towards replaceing the Spec API with KComponent API, I think just ending with these error messages is better for now, as it’s actionable for the dev and we’re not further investing in spec API.

---------

Reviewed By: adityasharat

Differential Revision: D53206074

fbshipit-source-id: 4c4fad8c08b74556983d1378bc43c6537a3d3658