Consider increased height when truncation spans are applied
Summary:
This diff aim at fixing a bug that happens when applying truncation spans also changes the final heigh of the calculated layout. This issue became apparent when adding tests for the new rich text component with support for composable spans. When testing the same scenario with the currently available rich-text component the two appeared consistent.

The bug was caused by the fact that we apply truncation spans after having laid out the text and verified that it doesn't fit the constraints. When this is the case we apply the truncation spans that in some cases (like the ones in the examples below) might impact the final measurements of the layout. In this diff we update the height value with the measured layout including the truncation text.

Reviewed By: pasqualeanatriello

Differential Revision:
D53306877

Privacy Context Container: L1152429

fbshipit-source-id: 5fa267b8730150fede2bcce0eac03123a82f8b4c