Fix the bug that `TextSpec` doesn't re-layout when `minimallyWide` is enabled
Summary:
# Context
This diff is to fix a bug that was identified from this oncall [post](https://fb.workplace.com/groups/litho.support/permalink/3693122427675765/), basically it's because we didn't re-layout text with the "minimal" size. The issue is hard to be noticed because we usually use the default text alignment(START) rather than `CENTER`.

This should also address the similar [issue](https://fb.workplace.com/groups/litho.support/permalink/2961919757462706/), which was posted few years ago.

`minimallyWide` was introduced by D13595632, which was basically trying to shrink the width of the text to make it take up as less space as possible.

Reviewed By: astreet

Differential Revision: D53041943

fbshipit-source-id: 44c4bee570b8215cb5b9ca79c372ad7f23064f3c