Fix weird layout behavior with minimallyWide
Summary: The purpose of this change is to fix the bug introduced by the previous diff, where we request a second layout inside `onBoundDefined` when taking `minimallyWide` into account. As we can see, the layout result has changed somehow, the last word would be moved to the next line with the same width, which is unexpected. I spent some time looking into this bug but wasn't able to figure out the root cause, the only way works for far is moving the second layout pass to `onMeasure` as well, as Andy suggested in the last diff.

Reviewed By: astreet

Differential Revision: D53404333

fbshipit-source-id: 90c2c8c2f25c9ac776cc40dcc74e9802f0278f97