Fix the bug of TextSpec where retrieving width of TextLayout incorrectly
Summary:
# Context

This diff is to correct the `minimallyWide` behavior of `TextSpec`. As we were using an incorrect API, `LayoutMeasureUtil.getWidth`, to get the max width of text layout. `getLineRight` doesn't always mean the width of each line, which depends on the text alignment, and using a different size constraints can lead to different layout result.

https://www.internalfb.com/code/fbsource/[9e996ce3eb7107a1eba6a921f2e40aeb384781db]/fbandroid/libraries/textlayoutbuilder/library/src/main/java/com/facebook/fbui/textlayoutbuilder/util/LayoutMeasureUtil.java?lines=43

To make it clear, I have a diagram to explain the problem:

 {F1453591596}

So we ended up with the second solution because we couldn't find a reliable way to make TextLayout generate exactly the same result with different size constraints.

Reviewed By: astreet

Differential Revision: D53440492

fbshipit-source-id: de7160e997f5a8030a0e9d983dffb04cd18d0ad2