Correcly calculate layout when applying truncation spans
Summary:
This diff follows-up on the original one D53306877 that was created to avoid truncation spans from causing layout to be cut. As in the pictures below that are using D53307069 to compares two versions of the Rich Text component, the one we currently have as part of the standard library and the new one that supports composable spans:

| Before | After |
| -- | -- |
| {F1423790792} | {F1423784358} |

The diff linked above was using the resulting height of the layout calculated after truncation and it overrode the previously calculated height.
This is not really possible in every circumstance and would need to be adjusted taking into account cap and baseline and other properties that were used to calculate the height before.
That changed caused regression and was reverted D53356716.

To fix the miscalculation, the truncation processing has been moved as a first step before calculating the preferred height.

Reviewed By: pasqualeanatriello

Differential Revision:
D53406823

Privacy Context Container: L1152429

fbshipit-source-id: 75f80d39bbd8c70ffde261b5322d3e3496a5353e