fix text auto truncation
Summary:
Force enabling the auto truncation MC for IG and FB in D53277172 resulted in changed truncation behavior for a couple of screens. Some changes were legit and some unexpected. This is unexpected for instance:

*L (auto-truncation on) | R (auto-truncation off)*

 {F1456695931}

while this is correct
 {F1456698194}
*See attached task for more examples*

I've made a couple of corrections to how we auto-truncate text in light of this.
1. We now check that the unbounded text layout's height fits the measured height from initial text layout. This piece is important to ensure that we don't go on and truncate unnecessarily. We should only truncate if the height exceeds the finite measured height
2. We no longer scan for only `AbsoluteSizeSpan`s as there may be other layout/metric affecting spans that we're not accounting for. The computation of line height is simpler and is based off of the font metrics from the original layout

With the above changes, the text truncation behavior should align more with iOS. I'll also update screenshots after this diff lands.

Reviewed By: rooju

Differential Revision: D53773679

fbshipit-source-id: a790b2ee9c25db9d9c54dff35ae3634a6632f4aa