setTree[Async] must not be debounced in `RenderState`
Summary:
State enqueues can only be honored if there is a `ResolveFunc` already set. To ensure that that is the case, all `setTree` calls must immediately set the resolve function on the `RenderState`. However, the eventual execution of the resolve function *may* be debounced in the layout executor if necessary. Unfortunately, I broke this silently in D52862109. Fortunately, this bug is only visible if there are sync and async operations in sequence

I only caught this because I was implementing async state updates and the test cases I had for async state updates failed

Reviewed By: adityasharat

Differential Revision: D53713067

fbshipit-source-id: 04f450043430422357a018c13d5300f3c4128a82