Crash for internal builds if we detect any host modification
Summary:
As we are attempting to ship host recycling safely in different surfaces, I'm adding a new behavior that will crash when an invalid modification is detected. This will be defined through an MC so that I can for now only filter for employees and test users.

The goal is just that I can get some visibility on some issues across our codebase.

Reviewed By: zielinskimz

Differential Revision: D53805320

fbshipit-source-id: 1a7de0592cdbbdd814799ac613a0e9019e03e134