reuse previous layout data for resolvable nodes
Summary:
To keep nested litho tree layout efficient we need the previous `LayoutState` as cache. To access the previous layout state, we will cache the layout data from the resolved node and read it on the next layout pass.

I put the caching behind a flag as I want to be able to compare memory impact across different experiment groups for Litho resolve

Reviewed By: adityasharat

Differential Revision: D53745589

fbshipit-source-id: 75b56b57e8f1072f650315c19f16045b47f80482