Add text outlines to Litho
Summary:
## Group of diffs
This diff and the next diff D53866998 add the ability to create enforced shadows combined with outlines around the text to make these texts visible over images, videos and other motley surfaces.

In the example below texts have 2dip outline and 1dip shadow:
{F1457678854}

## This diff
This diff adds support for the `outlineWidth` parameter to the `TextSpec` and `TextDrawable`.

TextDrawable should paint twice because stroke should not reduce the inner surface of the glyphs (i.e. it should stroke outside). This can be achieved by drawing glyphs twice - with stroke and with fill, covering the inner parts of the strokes.

## Next diff
Next diff adds outline support to the TextSpec.
Next diff adds support to the FDSText.

Reviewed By: pentiumao

Differential Revision: D53865149

fbshipit-source-id: b40b1c05189d65b3864a6502ab4d255a40435ed3