Log span type for spannable crash in TextInputSpec
Summary:
# Context

There's a crash in the Platypus public test (and likely also in Armadillo TLC test), where `SpannableStringBuilder` crashes sometimes when the device's auto correct dialog is shown from in-app edit texts, and the user selects a suggestion that would lead to the input exceeding the `EditText`'s max length. I double-checked the scenario where this crash happens via the sapienz videos in the attached crash.

# Fix

This has been fixed in Orca TV's composer before: [post](https://fb.workplace.com/groups/messengerAndroid/posts/961162387986383) | D26610843

We essentially need to provide a custom `Editable.Factory` implementation that absorbs the crash in `setSpan`. This will make it so the auto-correct selections will not work for the user if applying the suggestion would lead to a crash, but that is better than crashing.

# This diff

Since we are using Litho text inputs in `MsysThreadView`, we decided to fix this issue in `TextInputSpec` itself. This crash fix has been replicated in a few places across the codebase, so best to fix it at the root for Litho use-cases. Note that the `SafeEditableFactory` is set in `onCreateMountContent` because of limitations in the mount spec API.

**More on that:** Initially we tried setting/resetting the factory in `onMount`/`onUnmount` but since these steps are performed when the edit text expands to accommodate additional lines of text, we would keep re-setting the factory. This is a problem because setting a new factory resets the cursor location to the start of the edit text, so the experience of typing multiline messages would be completely broken.

# Latest update
We're no longer swallowing the `IndexOutOfBoundsException` for the `maxLength` issue since we suspect that doesn't cause most of the hits from the linked crash. D54237722 should fix most of the relevant crashes. Instead we're rethrowing the exception with some additional information so we have more to go on in case future fixes are needed.

Reviewed By: nicholeic

Differential Revision: D54034057

fbshipit-source-id: 1a4b8ef3c3e3ea187007905eb1941e0dfea86662