change approach for auto-truncation
Summary: I'm updating the approach for truncation to take into account individual line heights. The previous line height calculation was based off of font metrics which are always uniform. This doesn't always work because line height may vary depending on the presence of tall spans in a line or characters with ascenders and descenders.

Reviewed By: rooju

Differential Revision: D54698703

fbshipit-source-id: 5c99c0bd6e3096fe69ce9d301c07cbeea50d2b3b