truncate RTL text from the left
Summary:
While working on D54698703, I found out that RTL text wasn't being truncated correctly – we were overtruncating. Essentially, we set the ellipsis position close to the end of the line to ellipsize as though we were truncating LTR text. The correct thing to do for RTL text is to move the truncation point to the left most part of the line.

As far as I can tell, this has been a long-standing bug. I'm going to test this fix in conjunction with the auto-truncation fix.

Reviewed By: rooju

Differential Revision: D54709471

fbshipit-source-id: 3b6bbea39fcdba83930e79ffa65388bf7ebd693a