Fix crash when trying to request layout not from UI thread
Summary:
This fix should solve the [14ac7efdf29e3832e17baeb901084e2e mid](https://fburl.com/logview/8c1kyt92).
This error happens when requesting layout from a thread different from the thread originally used to create the view, which is the UI thread.

This diff gates the fix, it's turned off by default and should only affect FBLite users in the experiment we are planning to open to verify it's working.

Reviewed By: pengj

Differential Revision: D54753682

fbshipit-source-id: f26e5f4fa1ec86c27f2b865b4053c36d21747990