layout cache accepts any key type
Summary:
Will be necessary for diff upstack

I'm changing the layout cache to use *any* key type for the sake of permitting caching with keys that are not: `Node` or `int`. There are situations where creating a stable, globally unique int key is not feasible nor is using a `Node` as cache key possible

**What prompted this?**

We want to be able to cache a nested tree's layout state for subsequent layout computations but the layout cache accessed during layout is owned by Bloks. This means that other Bloks components could already have cached items by integer keys but because the id generating mechanism is not shared between Bloks and Litho, there is the possibility of key collisions.

This diff makes the node-keyed cache use any objects as keys. The overall behavior should not change

Let me know your thoughts

Reviewed By: adityasharat

Differential Revision: D55036892

fbshipit-source-id: 234db9ceb558114c18041d04acbca41bc4fa334f