Fix a crash from the Android framework's spell checker implementation
Summary:
# Context
We're currently seeing a crash in the TextInputSpec setting a span that's too long to handle for the text in the view. Previously, we added additional logging to determine what spans were causing this to crash. After adding this logging, we can see that by far the most commonly occurring span in the crash is `android.text.style.SuggestionRangeSpan`. I was able to repro this crash happening in the Android framework code, thanks to this [Google Issue](https://issuetracker.google.com/issues/314288203). If you misspell a word, tap the highlighted word to show spell check suggestions, and delete some characters before the spell check prompt shows up, the app will crash if that word was at the end of the string.

# Changes Made
To fix this, I update the EditableFactory that we already have in the component to take the minimum of the end int and the current length of the String. This way, if the Android system (or anyone else I guess) tries to set a span that goes past the length of the string, we will instead set it to the end of the string. Additionally, I have also limited the scope of this fix to SuggestionRangeSpan, since the developer should see the exception if any other class tries to set a span with invalid boundaries.

# Notes
* There are similar solutions for this crash that involve overriding hidden methods in the EditText to also take the minimum of the length and end. However when testing locally, I see that the system calls spannable.setSpan instead of TextView.setSpan_internal, meaning this case still crashes. You can see the call to spannable.setSpan in the Android framework code for [updateSuggestions](https://cs.android.com/android/platform/superproject/main/+/main:frameworks/base/core/java/android/widget/Editor.java;l=4366). This is why I updated the SpannableStringBuilder in the Editable Factory instead of adding those method overrides for our case.

Reviewed By: adityasharat

Differential Revision: D54969262

fbshipit-source-id: 25cf784b34a77047d0574f5487212b8a62328e60