only clear extension state if we're setting a different tree
Summary: I was testing NT transition examples when I found that transitions simply don't work. We're always calling `onBeforeSettingNewTree` on the `LithoRenderTreeView` and clearing visibility and transition extension states before any new remount. I believe we should only do this when we have a new tree i.e a tree with a different id from what is currently set on the view

Reviewed By: adityasharat

Differential Revision: D55127551

fbshipit-source-id: a3a0c4359d903acf2a1d56cf9ccc69dbc89ee138