Add occluded viewport width + height to VisibilityChangedEvent
Summary:
# Context

With the occlusion fix now launched for fb4a, in order for Merlin to properly calculate the viewport coverage in the same way, we need to provide the transformed (occluded) rect width and height of the viewport.

# This diff
Add the transformed viewport height + width to VisibilityChangeEvent so that consumers (Merlin) can use this to calculate coverage percentage including the occlusion fix.

Reviewed By: adityasharat

Differential Revision: D55091986

fbshipit-source-id: f541f8a151e98203dca20645ff0da9c6ee383791