Introduce frame identifier calculation to the rendering pipeline
Summary:
The earlier attempt to measure frame drops during animations was naive and did not take into account stalls on the background work (bind, resolve, layout) that we do during rendering.

This diff stack attempts to fix that.  There are two parts to this change:

**1. Frame Id (this diff)**
We introduce the concept of a "frame identifier", which is an integer that represents the frame in which the last update was requested.  The frame identifier is held by `RenderState` and updated when a new render is requested.  This frame id is then plumbed through to the pipeline to the point where we commit, so that at the point we commit, we know which update we are committing.  Commit listeners receive this.

**2. Animation perf (next diff)**
Each animation registers a commit listener when started.  This commit listener uses the frameId provided to the commit listener callback to then compute how many frames were lost since the last commit happens.

Differential Revision: D54906022

fbshipit-source-id: 7512e0c9ccd8237d1e7e14c95807203a7c79570c