OneShotPreDrawListener experiments in litho-widgets
Summary:
For all cases where a OnPreDrawListener was used that removed itself on first callback, replace with OneShotPreDrawListener from androidx, which makes sure the listener is cleaned up even if the window is detached before the callback is called.

Making sure the listener is cleaned up helped reduce ANRs in IG, so we are hoping to see more of the same from more such changes.

Reviewed By: zielinskimz

Differential Revision: D55761761

fbshipit-source-id: d19cb9e97c2fb200da571961ac71ee7b514a59e3