Simplify cache storage and retrieval
Summary:
Currently, we have a CachedValueKey that's made up of the actual inputs to the cache entry. This implies that whenever an input changes, we'll never find the key in the cache since the key hash is made up of said inputs. The bigger side effect is that, we're always writing new key/value to the cache every time there's a slight change in the inputs, which can be quite catastrophic to memory. And even worse is we never really prune the contents of the cache

This diff attempts to address all of these issues by simplifying what a cache key is, which is really just a combination of global key and cache index.
The inputs are now instead held in a CacheValue class, along with the actual cached value, so that they can be verified for whether/not they have changed since the last evaluation.

This way, cache key is always stable, and we can avoid polluting the cache with tons of data that will almost never get revisited
A changed input now results in simply replacing the existing value rather than adding a completely new entry

Reviewed By: colriot

Differential Revision: D56307165

fbshipit-source-id: 0503a8d4239fe323b9071508581349758f027f67