Add support for Style.pivotPercent
Summary:
Adds support for pivot points for transforms.

Because we need to recycle Views but resetPivot is not exposed until API 28, the strategy is to enforce that the View we apply this to implements SupportsPivotTransform. On API 28, the implementation of that interface keeps the pivot centered in the middle of the View after reset by using onSizeChanged.

Reviewed By: colriot

Differential Revision: D56568897

fbshipit-source-id: 4d8d42c94483f70ca0d796abef98284e6b628944