Allow focusing inline links with physical keyboard
Summary:
This diff leverages the changes proposed in this one D51871589 to enable text to be focusable when a clickable span is part of it. At the moment the test is set as not focusable causing the "tab" navigation via keyboard to skip text even when it is expected to be focusable due to the presence of a clickable span.

The Bloks changes to customise the appearance of clickable text will be separate from this change since they might require alignment between platforms.

Reviewed By: blavalla

Differential Revision: D55760551

fbshipit-source-id: d522ae1a93a1732b3b6800b54628a570dfa39934