Replace old usages of getResources #17
Summary:
**Problem**
When we have a java method `getX()`, then we can use it both as a method `.getX()` and as a field `.x` in kotlin. The problem is that when a java class is converted it the method should be either converted into a field or kept as a method, in this case only one usage remains valid.

Right now `ComponentContext` is in java and some kotlin callsites use method syntax. In order to be able to convert `ComponentContext` into kotlin we need to unify all existing usages.

**What this diff is doing**
```
abgs -f kt$ .getResources\(\) | awk -F':' '{print "/Users/ashcherba/"$1;}' | sort | uniq | xargs arc fastmod --accept-all ".getResources\(\)" ".resources"
```
Or in other words it just replaces all `.getResources()` calls with `.resources` in kotlin files.

**Is this safe?**
Technically, as long as everything compiles, it should be. This is just a replacement by a regular expression, so technically this stack can affect other classes (not only `ComponentContext`), but since kotlin does not allow to have both property and getter with the same name (`x` property and `getX()` method) we should not change semantics (as long as everything compiles, because if we have a different class with getter in kotlin, then this replacement will break compilation, because the property does not exist). If we are updating a usage of a java class, then it should not cause any problems, because both calls are allowed.

**Out of the scope**
We have a bunch of methods like `getString()` directly in `ComponentContext` that can be used without getting resources directly from context. This stack does not clean up things like that, it covers only existing syntax, not using proper APIs.

Reviewed By: pengj

Differential Revision: D56823450

fbshipit-source-id: 7784378859fc598080ff8daa20a11d18a337ac8c