Add configuration to RecyclerBinderConfiguration to control
Summary: In this diff I'm adding a configuration that determines whether a given `RecyclerCollectionComponent` will use a primitive recycler or not in the nested recycler. This will allow us to fine-grained specify if a given product spec uses the primitive recycler or not (which I do in the next diff of the stack).

Reviewed By: pentiumao

Differential Revision: D56702129

fbshipit-source-id: 79c8e22588932f07a177671a75ec67048c629550