Add ability to use fine-grained LithoViewAttributesExtensionState
Summary:
## Context

We have discovered a bug that mainly shows up whenever host recycling is enabled. In this bug we can see that the `ComponentHost` view attributes were not correctly cleaned up. As a consequence, you might see a host being recycled and still showing an incorrect background.

After some investigation I have nailed it down to the fact that our `TransitionExtensions` can acquire mount items, and avoid unmounting them until an animation ends (https://fburl.com/code/v8se58ad). By the time the animation ends, and we try to unmount the item, the call to get the view attributes was returning `null`, and therefore we were not re-setting them properly (https://fburl.com/code/vdv13489). This happened because a new `LayoutState` had been "mounted" and the `LithoViewAttributesExtensionState` had been reset and cleared (loosing information).

Below you can find an example of a host that was recycled in an invalid state:

 {F1507237951}

## Solution
The easiest approach so far is to actually change the implementation of the `LithoViewAttributesExtensionState` to make sure that it doesn't clear the `ViewAttributes` information of a given `RenderUnit` until this render unit is unmounted.

To allow for this, I'm creating a setup where we can swap between the current state implementation and the one I am creating here.

Reviewed By: zielinskimz

Differential Revision: D56933078

fbshipit-source-id: bde1c498f2200694298f04350163e2150d3dbf2a