Save flexlayout measureResult directly into Java result array, avoiding book keeping via JNI bridges.
Summary:
## Context

See context in D57101250.

## This diff
After introduction of "idx" in callbacks in D57101250,
move the measureResult to be set directly in Java during the measurements themselves.

As we measure (and we may measure multiple times!) and we get the result from the upper layers, save it directly in `LayoutOutput` object.

Do note that we still return the whole measureResult to JNI, for access the width/height/baseline float values.

As for the flexlayout native part, we provide a dummy measureResult value of "bool:true" which we basically ignore.

Differential Revision: D57127871

fbshipit-source-id: f5e69291895feb88bb9252019ec2134ea1998b9e