Add State Updater logic
Summary:
We have recently found a bug which relates to the fact that our ComponentTree is the StateUpdater.
 {F1502324297}
ComponentContext holds a reference to a StateUpdater which is implemented by the ComponentTree, and that will at the end delegate state update enqueuing logic to the TreeState.

The bug can be shown on lists whenever there is a listener that is created in the initial state callback and that can trigger state updates. In this scenario, what happens is that the listener is actually capturing the ComponentContext, and that its StateUpdater is a ComponentTree (let’s call it CT#1, because it is the first one mentioned in this example).
The flow of an update can be seen below:
 {F1502326319}

Eventually, the user can scroll away, and CT#1 will be released whenever we recycle the LithoView. However, if the user scrolls back to that Component, we will create a new CT (let’s call it CT#2) and pass it the old TreeState (we do this because we preserve State even when we create new CTs for the same Component).

This leads to a situation, where the new ComponentContext StateUpdater is pointing to CT#2. However, the Component Listener is still pointing to the old ComponentContext which delegates to CT#1. Therefore, if there is a state update coming from the Listener, it will try to process it in CT#1, which is already released (and therefore a no-op will happen).

 {F1502326661}

The solution is to add Facade State Updater that will holds a “configurable” state updater that we would pass into the Component Tree creation, so that it creates the ComponentContext with it.

	*Diagrams and description of the problem by fabiocarballo*

Reviewed By: fabiocarballo

Differential Revision: D56705985

fbshipit-source-id: 8724f52a94157a947da91c8767a3c8c9a9507c88