Implement transition with dependency logic
Summary:
Add implementation logic for transitions with dependencies.

This diff also contains an optimization to pre-calculate the transition on a background thread, so that it can be reused in cases where there hasn't been any layout state commit since the last evaluation. Otherwise, the transition is re-evaluated during mount time to capture all the latest dependency diffs

Reviewed By: adityasharat, astreet

Differential Revision: D56634283

fbshipit-source-id: 6e83fc507b5a7541d12223c7b849d41f970c9801