add option to disable recycling of the nested tree host
Summary:
The last run of the nested tree experiment was spiking JNI state animator FADs. We believe that this may be due to the host view being recycled.

This diff adds a flag to disable the recycling of the nested tree's host view. We'll run an experiment to understand if this helps with the crashes we've been seeing

Reviewed By: pasqualeanatriello

Differential Revision: D57353588

fbshipit-source-id: 54f9396265417effc52465a3601f05fc51053512