Add simple transitions to TransitionData
Summary:
Start the process to consolidate all transitions into the TransitionData class.  Transitions with dependency already use this class, so this diff focuses on migrating the existing transitions (without dependency).
At the end, we'd be able to completely get rid of Transition from the render pipeline, and channel everything through TransitionData

Reviewed By: adityasharat

Differential Revision: D57621169

fbshipit-source-id: 542d1d6746969697cb32c0e7d1aa47443f0d27ca