Add controller zoomable funcitonality
Summary:
Adding zoomable component that allows pinch to zoom/double tap to zoom functionality

Adding functionality, it is the same one as in bloks with a small difference that i kept controller out of the container and used viewBinder to bind the view with the controller. Due to that we have init() function where we are setting everything that requires the view.

Reviewed By: adityasharat, zielinskimz

Differential Revision: D57345505

fbshipit-source-id: 1db799dff8dba69c150ae9f6546afd5d3170559c