Fix crash regarding animator usage
Summary:
We have discovered one of our crashes is related to re-using the same animator in different RVs. This was because on `PrimitiveRecycler/CollectionRecycler` we were always re-using the same instance instead of recreating it as we do in `RecyclerCollectionComponentSpec/RecyclerSpec`

To tackle this, we are now supporting the same behavior that is shown in `RecyclerCollectionComponentSpec` regarding usage of default item animators. The idea is that we detect if the client chose the default item animator, and if so we propagate a new instance of it. This is because we can't re-use the same exact instance or it can lead to crashes (if ends up being used in two RVs at the same time; this was the case where I reproduced where I had a vertical list with a nested horizontal list).

I'm also adding a new configuration which allows this new system to be enabled in `LazyCollections` just to make sure this doesn't bring hell.

Reviewed By: zielinskimz

Differential Revision: D57898942

fbshipit-source-id: b1378ffde4e3977cba9cb7ff90d835f52fbf08c6