Fix the bug of IncrementalMount
Summary:
# Context

Since `getLocalVisibleRect` cannot return a correct value when view gets detached, which ends up with we're not able to do IM "correctly" in this special case.

- We run IM with an empty rect when LV gets detached, which means we expect all components that are not excluding from IM should be unmounted because they're invisible and not referenced by anyone.
- We run IM again when LV gets attached, and we could get the correct rect at this moment because LV has been attached, using this visible area to mount those components that are supposed to be mounted.

**Q: Why do we have to run IM again when getting re-attached?**
Since we've unmounted all possible items when detaching, if LithoView is attached again we need to attach component tree as well. As you can see from the code below, if there's no need to re-compute the layout we'll do `rebind`, which means we have to mount them before that!

https://www.internalfb.com/code/fbsource/[1e4d05d603bb0d30fc4c6ad908b49a4cadd71e9e]/fbandroid/libraries/components/litho-core/src/main/java/com/facebook/litho/ComponentTree.java?lines=729-733

**How does getLocalVisibleRect work?**

{F1643598406}

**Q: Why can't we use the callback of viewport manager, like [RecyclerBinder.viewportChanged](https://fburl.com/code/r0tosqi3)?**
Since this callback is being called after onAttach/onDetach happens, we never have the chance to trigger IncrementalMount correctly.

Reviewed By: adityasharat

Differential Revision: D57333975

fbshipit-source-id: 4f5261f28d3ca4a435fe3d7acf308c7d0ddbb729